Refactor deal-detail-view into single-responsibility units
Split the 1229-line monolith into focused files:
- hooks/use-deal-detail.ts: data fetching, Trustless/wallet hooks, role derivation
- components/deals/deal-milestone-actions.tsx: milestone card + handleAcceptOrder
- components/deals/deal-funding-panel.tsx: fund + extend-window UI and handlers
- components/deals/deal-delivery-panel.tsx: delivery confirmation dialog and handler
- components/deals/deal-detail-sidebar.tsx: return calculator, stakeholders, timeline
- components/deals/deal-detail-skeleton.tsx: loading skeleton
- components/deals/deal-detail-tabs.tsx: details/on-chain/documents tabs

deal-detail-view.tsx is now a 200-line orchestration shell.

Closes #62
